improving api II

+++ src/kernels/gelu.cu
@@ -1,7 +1,6 @@
 // GELU elementwise -- un thread por elemento. Usa to_float/from_float
 // de common.cuh para no repetir la logica de conversion de dtype.
-#include "kernels/common.cuh"
-
+#include "../../include/kernels/common.cuh"
 #ifndef KERNEL_DTYPE
 #define KERNEL_DTYPE float
 #endif
@@ -9,7 +8,7 @@
 using namespace yadrakova::kernels::common;
 
 extern "C" __global__ void gelu_kernel(
-    const KERNEL_DTYPE* input, KERNEL_DTYPE* output, int n)
+    const KERNEL_DTYPE* input, KERNEL_DTYPE* output, int64_t n)
 {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (idx >= n) return;
